refactor: 상품 Repository 계층 분리 및 컨트롤러 수정
