+++ product-service/src/main/java/com/monkey/productservice/infrastructure/persistence/ProductJpaRepository.java
@@ -1,13 +1,13 @@
-package com.monkey.productservice.domain.repository;
+package com.monkey.productservice.infrastructure.persistence;
 
 import com.monkey.productservice.domain.entity.ProductEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface ProductRepository {
-    ProductEntity save(ProductEntity productEntity);
+public interface ProductJpaRepository extends JpaRepository<ProductEntity, UUID> {
     Optional<ProductEntity> findByIdAndIsDeletedFalse(UUID id);
     List<ProductEntity> findAllByIsDeletedFalse();
 }
